requirements.txt
flask
flask-cors
opencv-python
tensorflow
numpy
Pillow
